chore: Allow format `<class name>.cmd(args)` fmt
If a user enters a command like `User.all()` then
consider it a valid command and execute it.
The valid `cmd` options are:
    - all
    - update
    - show
    - count
    - destroy

Various commands take in various arguments.
As of now only the commands and class name part are done.
`args` part is yet to be done as it requires more parsing.

For now the `line` argument is manipulated before passing it on.
This saves on having to call the "right" command and lets
the previous implementation to do it (thanks to previous).

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>