chore: Add support for args
Args are now parsed as properly as possible.
There is still a problem with how numbers are written.
Instead of the age (an int) being written as so, it
is instead written as a string.

This has to do with how the cli parsing is done in general and
nothing to do with how the `precmd` is handling the args.
This behaviuor may be fixed in future iterations (if any).

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>